remove circular dep between item effects and pokemon entity
